only premake things with existing pyodide package directories

+++ get_built_pyodide.sh
@@ -8,10 +8,11 @@ import json,os,shutil
 with open("pyodide/build/packages.json","r") as f:
   plist=json.load(f)
   for package,info in plist['packages'].items():
-    print(f"Making dist for {package}")
-    os.makedirs(f"pyodide/packages/{info['name']}/build/dist",exist_ok=True)
-    shutil.copyfile(f"pyodide/build/{info['file_name']}",f"pyodide/packages/{info['name']}/build/dist/{info['file_name']}")
-    open(f"pyodide/packages/{info['name']}/build/.packaged", 'w'). close()
+    if os.path.exists(f"pyodide/packages/{info['name']}/meta.yaml"):
+        print(f"Making dist for {package}")
+        os.makedirs(f"pyodide/packages/{info['name']}/build/dist",exist_ok=True)
+        shutil.copyfile(f"pyodide/build/{info['file_name']}",f"pyodide/packages/{info['name']}/build/dist/{info['file_name']}")
+        open(f"pyodide/packages/{info['name']}/build/.packaged", 'w'). close()
 EOF
 
 # numpy needs building into packages/.artifacts in pyodide for pyarrow to build -
